Refactor backend
Second major refactor of the backend. This is a first step towards
fixing some of the UI blocking issues. Now we use a single-producer,
multiple consumer pattern when dealing with loading/downloading works.
The next step is to have multiple producers as well to deal with retrying
when rate limited in loading users, series, etc.

Misc:
  * Bumped maximum concurrency limit from 20 -> 50
  * Added indicator for where file is being downloaded to
  * Always show the work ID even when not loaded